pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "consent-mcp"
version = "0.1.0"
description = "A Model Context Protocol server for AI agent consent management"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Consent MCP Contributors" }
]
keywords = ["mcp", "consent", "ai", "agents", "twilio", "sendgrid"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "twilio>=9.0.0",
    "sendgrid>=6.11.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
    "fastapi>=0.109.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "pre-commit>=3.6.0",
    "testcontainers[postgres]>=4.0.0",
]

[project.scripts]
consent-mcp = "consent_mcp.mcp.v1.server:main"

[project.urls]
Homepage = "https://github.com/sairajm/consent-mcp"
Repository = "https://github.com/sairajm/consent-mcp"
Issues = "https://github.com/sairajm/consent-mcp/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/consent_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["consent_mcp"]
